requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==1.97.1
python-dotenv==1.0.0
